requirements.txt
flask-restful
logmetrics-sdk